azuriranje stanja namirnice
dinamicko dohvacanje podataka za modal.

+++ website/models.py
@@ -1,4 +1,5 @@
 # storage of database models
+from decimal import Decimal
 from pony.orm import Database, Required, Optional
 
 db = Database()
@@ -6,5 +7,7 @@ db = Database()
 
 class Namirnica(db.Entity):
     ime_namirnice = Required(str, unique=True)
-    stanje_namirnice = Required(float)
+    # Decimal jer se onda float racunanje vrsi optimalno;
+    # sa decimal 1.1 + 2.2 nece biti 3.3000000000000003 nego 3.3
+    stanje_namirnice = Required(Decimal)
     mjerna_jedinica = Required(str)
